+++ PixelsorterTests/SorterTests/SorterTestHelpers.cs
@@ -6,10 +6,10 @@ namespace Pixelsorter.Tests.SorterTests
     {
         // HSL pixel definitions (H: 0-360, S: 0-1, L: 0-1)
         // Saturation ordering: Gray (0) < LowSaturation (0.3) < MidSaturation (0.5) < HighSaturation (1.0)
-        public static readonly float[] Gray =           [0f, 0.0f, 0.5f];
-        public static readonly float[] LowSaturation =  [0f, 0.3f, 0.7f];
-        public static readonly float[] MidSaturation =  [0f, 0.5f, 0.6f];
-        public static readonly float[] HighSaturation = [0f, 1.0f, 0.4f];
+        public static readonly Half[] Gray =           [(Half)0, (Half)0, (Half)0.5];
+        public static readonly Half[] LowSaturation =  [(Half)0, (Half)0.3, (Half)0.7];
+        public static readonly Half[] MidSaturation =  [(Half)0, (Half)0.5, (Half)0.6];
+        public static readonly Half[] HighSaturation = [(Half)0, (Half)1.0, (Half)0.4];
 
         public static NDArray CreateUnsortedImageData()
         {
